+++ src/plugins/dashboard/public/application/embeddable/dnd_kit_grid/__stories__/draggable.stories.tsx
@@ -5,23 +5,18 @@
  * in compliance with, at your election, the Elastic License 2.0 or the Server
  * Side Public License, v 1.
  */
-import React, { useState, useCallback, useMemo } from 'react';
-import { DndContext, DragMoveEvent, DragOverlay } from '@dnd-kit/core';
-import { restrictToWindowEdges } from '@dnd-kit/modifiers';
-import { css } from '@emotion/react';
-
-import { Draggable } from '../components/container_components/draggable';
-import { PanelState } from '../types';
+import React from 'react';
+import { DroppableGrid } from '../components/container_components/droppable_grid';
 
 export default {
   title: 'POC - dnd-kit/Draggable',
-  component: Draggable,
+  component: 'Draggable',
   description: 'POC of new grid layout system',
   argTypes: {},
 };
 
-export const SnapToGrid = () => {
-  const [gridState, setGridState] = useState<Record<string, PanelState>>({
+export const BasicSnapToGrid = () => {
+  const gridState = {
     ['panel1']: {
       id: 'panel1',
       pos: { x: 0, y: 0 },
@@ -30,82 +25,7 @@ export const SnapToGrid = () => {
       id: 'panel2',
       pos: { x: 50, y: 50 },
     },
-  });
-  const [draggingId, setDraggingId] = useState<string | undefined>(undefined);
-
-  const handleDragStart = (event: DragMoveEvent) => {
-    setDraggingId(event.active.id);
   };
 
-  const handleDragEnd = (event: DragMoveEvent) => {
-    setDraggingId(undefined);
-    setElementPos(event);
-  };
-
-  const setElementPos = (event: DragMoveEvent) => {
-    const id = event.active.id;
-    const newPanelState = gridState[id];
-
-    newPanelState.pos.x = Math.max(newPanelState.pos.x + event.delta.x, 0);
-    newPanelState.pos.y = Math.max(newPanelState.pos.y + event.delta.y, 0);
-
-    setGridState({ ...gridState, [id]: newPanelState });
-  };
-
-  const guttersize = 5;
-  const columns = 12;
-  const maxRow = 30;
-  const CELL_HEIGHT = 26;
-
-  const gridStyles = useMemo(
-    () =>
-      css`
-        height: 100%;
-        width: 100%;
-        display: grid;
-        gap: ${guttersize}px;
-        grid-template-columns: repeat(
-          ${columns},
-          calc((${100}% - ${guttersize * (columns - 1)}px) / ${columns})
-        );
-        grid-template-rows: repeat(${maxRow}, ${CELL_HEIGHT}px);
-        justify-items: stretch;
-      `,
-    [guttersize, columns, maxRow]
-  );
-
-  return (
-    <div id="gridContainer" className="dshGrid dshLayout--editing" css={gridStyles}>
-      <DndContext
-        onDragStart={handleDragStart}
-        onDragEnd={handleDragEnd}
-        modifiers={[restrictToWindowEdges]}
-      >
-        {Object.keys(gridState).map((id) => (
-          <Draggable id={id} position={gridState[id].pos}>
-            <div
-              style={{
-                height: '100%',
-                width: '100%',
-                border: 'dashed 1px red',
-                background: 'pink',
-                padding: '10px',
-              }}
-            >
-              I am {id}
-            </div>
-          </Draggable>
-        ))}
-        <DragOverlay dropAnimation={null}>
-          {draggingId ? (
-            <div
-              style={{ height: '100%', width: '100%', background: 'lightblue', padding: '10px' }}
-            >
-              Dragging {draggingId}
-            </div>
-          ) : null}
-        </DragOverlay>
-      </DndContext>
-    </div>
-  );
+  return <DroppableGrid startingState={gridState} />;
 };
